Create a RequestRatelimiter and use it instead of BaseHandler.ratelimit.
